+++ components/Articles/ArticlesCreate/ArticlesCreateButtons.vue
@@ -44,16 +44,8 @@
 </template>
 
 <script setup lang="ts">
-/* import { Buffer } from "buffer"; */
-import { EncodeObject, MsgCreatePostEncodeObject, MsgSaveProfileEncodeObject } from "@desmoslabs/desmjs";
-import Long from "long";
-import { v4 as uuidv4 } from "uuid";
 import { useAccountStore } from "~~/core/store/AccountStore";
 import { useDraftStore } from "~~/core/store/DraftStore";
-/* import { useBackendStore } from "~~/core/store/BackendStore";
-import { usePostStore } from "~~/core/store/PostStore";
-import { useUserStore } from "~~/core/store/UserStore"; */
-import { useDesmosStore } from "~~/core/store/DesmosStore";
 import { usePostStore } from "~~/core/store/PostStore";
 
 const isPublishing = ref(false);
@@ -86,129 +78,12 @@ function saveDraft () {
 }
 
 async function publish () {
-  const { $useTransaction, $useIpfs } = useNuxtApp();
   isPublishing.value = true;
-
-  // check if the user has a sectionId
-  if (useAccountStore().sectionId <= 0) {
-    // if not, create one
-    if (useAccountStore().sectionId === -10) {
-      await useAccountStore().getUserSection(true);
-      // TODO: handle failure?
-    }
-    let attempt = 0;
-    // wait 10 attempts to create a sectionId, create the group, and user join it
-    while (attempt < 10 && useAccountStore().sectionId <= 0) {
-      // update the sectionId from the user endpoint
-      await useAccountStore().getUserSection();
-
-      // if found, break the loop
-      if (useAccountStore().sectionId > 0) {
-        break;
-      }
-      // otherwise, wait 5s and try again
-      await new Promise(resolve => setTimeout(resolve, 5000));
-      attempt++;
-    }
-  }
-  if (useAccountStore().sectionId <= 0) {
-    isPublishing.value = false;
-    throw new Error("Could not get section id");
-  }
-
-  const draftStore = useDraftStore();
-  const extId = uuidv4();
-
-  const msgCreatePost: MsgCreatePostEncodeObject = {
-    typeUrl: "/desmos.posts.v2.MsgCreatePost",
-    value: {
-      subspaceId: Long.fromNumber(useDesmosStore().subspaceId),
-      externalId: extId,
-      attachments: [],
-      author: useAccountStore().address,
-      text: draftStore.title,
-      sectionId: useAccountStore().sectionId,
-      tags: useDraftStore().tags.map(tag => tag.content.value),
-      conversationId: Long.fromNumber(0),
-      referencedPosts: [],
-      replySettings: 1
-    }
-  };
-
-  // craft a custom object for IPFS
-  const ipfsPost: any = {
-    ...msgCreatePost.value,
-    subtitle: draftStore.subtitle,
-    content: draftStore.content
-  };
-
-  // upload the post to IPFS (without CID attachment), get the returned CID
-  const postCid = await $useIpfs().uploadPost(JSON.stringify(ipfsPost));
-
-  const postIpfsUrl = `${$useIpfs().gateway}${postCid}`;
-  console.log(postIpfsUrl);
-
-  const ipfsEntityUrl = {
-    displayUrl: "IPFS",
-    start: Long.fromNumber(0),
-    end: Long.fromNumber(1),
-    url: postIpfsUrl
-  };
-
-  // attach the CID to the Post as an entity
-  msgCreatePost.value.entities = {
-    hashtags: [],
-    mentions: [],
-    urls: [ipfsEntityUrl]
-  };
-
-  /* $useTransaction().push(msgCreatePost);
-  const signedBytes = await $useTransaction().execute(); */
-
-  // if is a new user and has no profile, create one with the randomly generated username
-  const msgs: EncodeObject[] = [];
-  const msgsDetails: any[] = [];
-  if (useAccountStore().isNewProfile) {
-    const msgSaveProfile: MsgSaveProfileEncodeObject = {
-      typeUrl: "/desmos.profiles.v3.MsgSaveProfile",
-      value: {
-        dtag: useAccountStore().profile.dtag,
-        nickname: useAccountStore().profile.nickname,
-        bio: useAccountStore().profile.bio,
-        profilePicture: useAccountStore().profile.pictures.profile,
-        coverPicture: useAccountStore().profile.pictures.cover,
-        creator: useAccountStore().address
-      }
-    };
-    msgs.push(msgSaveProfile);
-    msgsDetails.push({
-      dtag: useAccountStore().profile.dtag,
-      nickname: useAccountStore().profile.nickname,
-      bio: useAccountStore().profile.bio,
-      profile: useAccountStore().profile.pictures.profile,
-      cover: useAccountStore().profile.pictures.cover,
-      scriptaOp: "MsgSaveProfile"
-    });
-  }
-  // push the post message
-  msgs.push(msgCreatePost);
-  msgsDetails.push({
-    externalId: msgCreatePost.value.externalId,
-    author: msgCreatePost.value.author,
-    sectionId: msgCreatePost.value.sectionId,
-    text: msgCreatePost.value.text,
-    tags: msgCreatePost.value.tags,
-    subtitle: useDraftStore().subtitle,
-    content: useDraftStore().content,
-    entities: JSON.stringify(msgCreatePost.value.entities),
-    scriptaOp: "MsgCreatePost"
-  });
-
-  const success = await $useTransaction().directTx(msgs, msgsDetails);
+  const success = await usePostStore().savePost();
   if (success) {
     await usePostStore().updateUserPosts();
     useDraftStore().$reset();
-    useRouter().push(`/@${useAccountStore().profile.dtag}/${extId}`);
+    useRouter().push(`/@${useAccountStore().profile.dtag}/${useDraftStore().externalId}`);
   }
   isPublishing.value = false;
 }
